% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmercari.R
\name{scale_color_mercari}
\alias{scale_color_mercari}
\title{Mercari geom}
\usage{
scale_color_mercari()
}
\description{
most of the code is from emoGG https://github.com/dill/emoGG
}
\author{
Megan McMahon
}
